+++ clandestined/test/test_clandestined.py
@@ -1,13 +1,10 @@
 
 import unittest
 
-from test_rendezvous_hash import RendezvousHashTestCase
-from test_rendezvous_hash import RendezvousHashIntegrationTestCase
-
-from test_cluster import ClusterTestCase
-from test_cluster import ClusterIntegrationTestCase
-
-from test_docs import DocTestCase
+from test_docs import *
+from test_cluster import *
+from test_collision import *
+from test_rendezvous_hash import *
 
 if __name__ == '__main__':
     unittest.main()